requirement.txt
Flask
bcrypt
pyjwt
mysql-connector-python